v3.0.0a12-5
Codebasechanges
- added argon2 wrapper for reusability
- added base64 function class for reusability
- used password generator class for all password gen related functions (no repetition)
