+++ app/controllers/TextEncryptionController.js
@@ -88,8 +88,4 @@ export class TextEncryptionController {
           ElementHandler.crossToArrows();
         }
       }
-
-      delay(ms) {
-        return new Promise((resolve) => setTimeout(resolve, ms));
-      }
   }
