+++ app/insertVersion.js
@@ -1,3 +1,3 @@
-appVersion = "3.0.0a12-4";
+appVersion = "3.0.0a12-5";
 
 $(".version").text(appVersion);
